Separate development protocol playbook
